Grabbing user ocation completed
