+++ Experiences/Experiences/Application/Extensions.swift
@@ -1,0 +1,29 @@
+//
+//  Extensions.swift
+//  Experiences
+//
+//  Created by Wyatt Harrell on 5/8/20.
+//  Copyright © 2020 Wyatt Harrell. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+extension UIViewController {
+    func hideKeyboardWhenViewTapped() {
+        let tap: UITapGestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(UIViewController.dismissKeyboard))
+        tap.cancelsTouchesInView = false
+        view.addGestureRecognizer(tap)
+    }
+    @objc func dismissKeyboard() {
+        view.endEditing(true)
+    }
+}
+
+extension UITextField {
+    func setLeftPaddingPoints(_ amount:CGFloat){
+        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: amount, height: self.frame.size.height))
+        self.leftView = paddingView
+        self.leftViewMode = .always
+    }
+}
