+++ Experiences/Experiences/Application/Info.plist
@@ -64,5 +64,7 @@
 	<string>$(PRODUCT_NAME) needs access to your camera to record video.</string>
 	<key>NSMicrophoneUsageDescription</key>
 	<string>$(PRODUCT_NAME) needs the microphone to create audio recordings.</string>
+	<key>NSLocationWhenInUseUsageDescription</key>
+	<string>$(PRODUCT_NAME) needs access to your location to save videos.</string>
 </dict>
 </plist>
